refactor(react): move i18n provider hooks to files
